+++ app/controllers/editions_controller.rb
@@ -12,7 +12,7 @@ class EditionsController < ApplicationController
     @selections = @edition.selections_with_categories.includes(ratings: :critic).order("films.title")
     @selections_by_category = @selections.group_by do |selection|
                                 selection.film.categories.first
-                              end.sort_by { |category, _selections| category&.name }
+                              end.sort_by { |category, _selections| category&.position }
     @critics = @selections.map(&:ratings).flatten.compact.map(&:critic).uniq.sort_by(&:last_name).reject do |critic|
       hidden_critics&.include?(critic.id.to_s)
     end