+++ config/routes.rb
@@ -22,6 +22,10 @@ Rails.application.routes.draw do
       resources :editions do
         get :search_for_film_to_add, to: "films#search_for_film_to_add_to_edition"
 
+        resources :categories, only: [:index] do
+          patch :reorder, on: :member
+        end
+
         resources :selections do
           get :csv, on: :collection
           post :import, on: :collection