+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2024-05-04
+- Category ordering
+  - A drag-and-drop view for defining the order categories will be displayed in grid
+  - Only Admins can see the Category view in the admin menu for an Edition
+
 ## 2024-05-01
 - Added ability to filter out Critics from the grid and reset the filters
 - Added director name to Film column